fixes
* pragma once
* float typecasting

+++ game/codegiraffe/v2.h
@@ -625,8 +625,8 @@ struct V2 {
 	static V2<TYPE> randomUnitVector() {
 		V2<TYPE> r;
 		do {
-			r.x = (rand() & 0x7fffffff) - (rand() & 0x7fffffff);
-			r.y = (rand() & 0x7fffffff) - (rand() & 0x7fffffff);
+			r.x = (float)((rand() & 0x7fffffff) - (rand() & 0x7fffffff));
+			r.y = (float)((rand() & 0x7fffffff) - (rand() & 0x7fffffff));
 		} while (r.isZero());
 		return r.normal();
 	}
